Move lower_get_args to ogc
This is the only place where
the function is used, and
moving it makes it possible
to import http_utils without
bringing in flask.

Refs #1549
