+++ datacube_ows/http_utils.py
@@ -8,8 +8,6 @@ from __future__ import annotations
 import json
 from urllib.parse import urlparse
 
-from flask import request
-
 TYPE_CHECKING = False
 if TYPE_CHECKING:
     from collections.abc import Mapping
@@ -86,22 +84,6 @@ def cache_control_headers(max_age: int) -> dict[str, str]:
     if max_age <= 0:
         return {"cache-control": "no-cache"}
     return {"cache-control": f"max-age={max_age}"}
-
-
-def lower_get_args() -> dict[str, str | None]:
-    """
-    Return Flask request arguments, with argument names converted to lower case.
-
-    Get parameters in WMS are case-insensitive, and intended to be single use.
-    Spec does not specify which instance should be used if a parameter is provided more than once.
-    This function uses the LAST instance.
-    """
-    d: dict[str, str | None] = {}
-    for k in request.args:
-        kl = k.lower()
-        for v in request.args.getlist(k):
-            d[kl] = v
-    return d
 
 
 def json_response(result: CFG_DICT, cfg: OWSConfig) -> FlaskResponse:
